requirements.txt
keras==2.15.0
numpy==1.26.4
pandas==2.2.1
plotly==5.19.0
pydeck==0.8.1b0
Requests==2.31.0
scikit_learn==1.3.2
snowflake_connector_python==3.7.1
snowflake_snowpark_python==1.12.1
streamlit==1.30.0
streamlit_lottie==0.0.5
streamlit_option_menu==0.3.12
tensorflow==2.15.0
